pyproject.toml
[tool.coverage.run]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 60
skip_empty = true

[tool.coverage.html]
directory = ".tests_coverage"

[tool.mypy]
check_untyped_defs = "true"
disallow_any_generics = "true"
disallow_incomplete_defs = "true"
disallow_untyped_defs = "true"
no_implicit_optional = "true"

[tool.poetry]
name = "qe-metrics"
version = "1.0.0"
description = "Report QE metrics"
authors = ["Caleb Evans <caevans@redhat.com>"]
readme = "README.md"
license = "Apche-2.0"
homepage = "https://github.com/RedHatQE/qe-metrics"
repository = "https://github.com/RedHatQE/qe-metrics"
packages = [{ include = "qe_metrics" }]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[tool.poetry.urls]
Download = "https://github.com/RedHatQE/qe-metrics"
"Bug Tracker" = "https://github.com/RedHatQE/qe-metrics/issues"

[tool.poetry.dependencies]
# 3.12 to be added following completion of https://github.com/ponyorm/pony/issues/703
python = ">=3.8,<3.12"
click = "^8.1.7"
jira = "^3.5.2"
jinja2 = "^3.1.2"
python-simple-logger = "^1.0.8"
psycopg2-binary = "^2.9.9"
pony = "^0.7.17"
pyaml-env = "^1.2.1"
ipdb = "^0.13.13"
pyhelper-utils = "^0.0.9"

[tool.poetry.group.dev.dependencies]
tox = "^4.11.4"
pre-commit = "*"
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.12.0"
ipython = "*"

[tool.poetry.scripts]
qe-metrics = "qe_metrics.cli:cli_entrypoint"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]
